return filtered and unfiltered price type options
